[refactor] feed / activity
