+++ frappe/hooks.py
@@ -35,18 +35,24 @@ before_tests = "frappe.utils.install.before_tests"
 
 website_generators = ["Web Page", "Blog Post", "Website Group", "Blog Category", "Web Form"]
 
+# login
+
+on_session_creation = "frappe.desk.doctype.feed.feed.login_feed"
+
 # permissions
 
 permission_query_conditions = {
 	"Event": "frappe.desk.doctype.event.event.get_permission_query_conditions",
 	"ToDo": "frappe.desk.doctype.todo.todo.get_permission_query_conditions",
-	"User": "frappe.core.doctype.user.user.get_permission_query_conditions"
+	"User": "frappe.core.doctype.user.user.get_permission_query_conditions",
+	"Feed": "frappe.desk.doctype.feed.feed.get_permission_query_conditions"
 }
 
 has_permission = {
 	"Event": "frappe.desk.doctype.event.event.has_permission",
 	"ToDo": "frappe.desk.doctype.todo.todo.has_permission",
-	"User": "frappe.core.doctype.user.user.has_permission"
+	"User": "frappe.core.doctype.user.user.has_permission",
+	"Feed": "frappe.desk.doctype.feed.feed.has_permission"
 }
 
 doc_events = {
@@ -55,10 +61,14 @@ doc_events = {
 		"validate": "frappe.email.doctype.email_alert.email_alert.trigger_email_alerts",
 		"on_update": [
 			"frappe.core.doctype.notification_count.notification_count.clear_doctype_notifications",
-			"frappe.email.doctype.email_alert.email_alert.trigger_email_alerts"
+			"frappe.email.doctype.email_alert.email_alert.trigger_email_alerts",
+			"frappe.desk.doctype.feed.feed.update_feed"
 		],
 		"after_rename": "frappe.core.doctype.notification_count.notification_count.clear_doctype_notifications",
-		"on_submit": "frappe.email.doctype.email_alert.email_alert.trigger_email_alerts",
+		"on_submit": [
+			"frappe.email.doctype.email_alert.email_alert.trigger_email_alerts",
+			"frappe.desk.doctype.feed.feed.update_feed"
+		],
 		"on_cancel": [
 			"frappe.core.doctype.notification_count.notification_count.clear_doctype_notifications",
 			"frappe.email.doctype.email_alert.email_alert.trigger_email_alerts"
